RAG_System_Example/requirements.txt
sentence-transformers 
faiss-cpu 
transformers 
langchain 
beautifulsoup4 
requests
